fix(dashboards): show the author's chart draft inside dashboard tiles
